requirements.txt
numpy
matplotlib
pillow
gym
imageio
pyperplan
networkx
z3-solver
jinja2
graphviz
ipython
ipykernel
tqdm
